requirements.txt
streamlit
langchain
PyPDF2
langchain-openai
langchain-core
langchain-community
python-dotenv
uuid